backend/requirements.txt
django
django-cors-headers
psycopg2-binary
Pillow
djangorestframework-simplejwt
djangorestframework
spacy
